Bug FIX 1636

+++ src/main/java/clock/Game.java
@@ -56,7 +56,7 @@ public class Game extends JPanel implements Mode, Runnable, Serializable {
                 dinosaur.dead(true);
                 gameState = GAME_OVER_STATE;
             }
-            dinosaur.setSpeedX((int)Math.min((floatSpeed * floatSpeed), floatSpeed));
+            dinosaur.setSpeedX((int)Math.min((double)(floatSpeed * floatSpeed), floatSpeed));
         }
 
     }
